refactor(routes): extract /plugins into plugins.js (god-file decomposition)
Route contract unchanged (284 routes), 0 /plugins left inline, vitest 38 passed 268 passed . Lucy (GLM-5.2), driver-gated.
